Add new fields row and col to the sudoku struct
These new fields track the last position of the method solve so it is
more efficient and faster now.

A new flag has been added to control the refresh rate of the output when
verbose mode is enabled.

The variables y and x have been renamed to row and col to make it more
readable.

The function new has been changed to make it more simple and readable.

A new variable screenCap has been created to remove a magic number.

Some other minor changes in comments and variable's names.

The license has been changed.