Modify the parse function
The parsing is done now using the package bufio instead of csv. While
the code of the function hasn't changed a lot, it allows now to read
from stdin as well. The code of this function has been moved to main.go
and the file parser.go has been deleted.

Some other minor changes had to be done in order for this to work
properly. For example, a new flag to change the delimiter has been
added since using "," was quite painful when introducing the
sudokus via stdin.

The value of *fps is now passed to the sudoku struct only if verbose is
enabled.

New .csv examples have been addeed to show different kind of inputs
using different delimiters and chars to represent the empty cells.

The README file has been modified to reflect the new changes and ways to
use the program.